Update from master (#717)
* Update images (#686)

Ensure that standard Win2012 machine has 4.6.2
Patch testimpact 2016 machine to see whether it fixes occasional bugchecks shortly after startup

* Update image list (#688)

* Update image list

Remove Nano Exclusion (now just covered by Helix)
Disable Win10, which is a non-functional 2016 TP5 image

* Always retry on NuGet cache corruptions to reduce its impact (#685)

* Always retry on NuGet cache corruptions to reduce its impact

Always retry on NuGet cache corruptions to reduce its impact. until it's
root caused and fixed.

* Always retry on NuGet cache corruptions to reduce its impact

Always retry on NuGet cache corruptions to reduce its impact. until it's
root caused and fixed.

* Adjust the tabs

Adjust the tabs

* Address review comments.

Address review comments.

* Add repro job functionality

Add repro job functionality

* Remove project-system's dev15.1.x branch (#690)

This branch no longer exists.

* Revert "Add CoreCLR branch dev/local-gc to CI (#663)" (#692)

This reverts commit bd42888661fed0dbf6ea83f6fe1400aca068ada2.

* Add MSBuild vs15.3 branch (#693)

* Revert to 20170303-elevated (#695)

This is just to unblock until elevation issue on 20170531 is fixed

* Adding .NET 4.6.2 SDK to win2012 images

* Enable CI for features/non-trailing (#698)

* Add dotnet/roslyn-analyzers dev15.3 branch (#702)

* Adding the 'Win2016 + VS2017.3 Preview 2' image (#703)

* Add retry logic for 'problem on deletion' xunit failure (#697)

* Update docker run hello-world on ubuntu1404-20160211-1 (#707)

* Add release/2.0.0-preview2 branches of CLI and SDK (#714)

+++ management/nodes/startupScripts/ubuntu1404-20160211-1.sh
@@ -1,5 +1,10 @@
 # Restart docker since now the mnt drive is writeable
 service docker restart
+docker run hello-world
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
 
 # Remove the file that allows dotnet-bot paswordless, nointeractive sudoing (since we are done sudoing now).
 rm /etc/sudoers.d/zzz-dotnet-bot
